数据更新

+++ community/io/src/main/java/org/neo4j/io/pagecache/tracing/EvictionEvent.java
@@ -1,6 +1,6 @@
 /*
- * Copyright (c) 2002-2017 "Neo Technology,"
- * Network Engine for Objects in Lund AB [http://neotechnology.com]
+ * Copyright (c) 2002-2020 "Neo4j,"
+ * Neo4j Sweden AB [http://neo4j.com]
  *
  * This file is part of Neo4j.
  *
@@ -55,7 +55,7 @@ public interface EvictionEvent extends AutoCloseablePageCacheTracerEvent
         }
 
         @Override
-        public void setCachePageId( int cachePageId )
+        public void setCachePageId( long cachePageId )
         {
         }
 
@@ -89,5 +89,5 @@ public interface EvictionEvent extends AutoCloseablePageCacheTracerEvent
     /**
      * The cache page id of the evicted page.
      */
-    void setCachePageId( int cachePageId );
+    void setCachePageId( long cachePageId );
 }
